Title: 실수 받아 그대로 출력, Time: 150ms, Memory: 12MB, Status: Passed - Codetree

+++ 251112/실수 받아 그대로 출력/enter-real-value-and-print.java
@@ -5,7 +5,7 @@ public class Main {
        
        Scanner sc = new Scanner(System.in);
 
-       int n = sc.nextDouble();
+       double n = sc.nextDouble();
 
        System.out.printf("%.2f", n);
     }
